Move HttpClientFactory to common to expose to other componenets
